Refactored explorer tree generation. Fixes IE long tree building.
Previously a lot of <script> nodes were injected into the html document, which called
tree building functions refering to the nodes in the document. It was happening as
the document was loading, hence there was a potential problem of slowing down things
on big documents, because page loading had to be interrupted each time there was a call
to the building of the tree element. While most browsers optimized it out(I guess they
only start executing scripts after the whole page is loaded) IE was really slow for big
documents. The solution was to wait until the page loads, then traverse the DOM and
extract useful nodes and build tree items.